refactor: move the different compressions into their own implementations
